flashcard/back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = "Backend's template"
authors = ["James <me@j4m3s.eu>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework-filters = "^0.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
